requirements.txt
watchdog==3.0.0
google-generativeai
python-dotenv